chore: move errorCardProperties.js and findErrorMessageIds.js to utils/
