+++ app/component/ItinerarySummaryListContainer/components/ItinerarySummaryMessage.js
@@ -5,10 +5,10 @@ import { matchShape } from 'found';
 import LocationStateShape from '../../../prop-types/LocationStateShape';
 import LocationShape from '../../../prop-types/LocationShape';
 import ErrorShape from '../../../prop-types/ErrorShape';
-import findErrorMessageIds from './findErrorMessageIds';
 import RoutingErrorShape from '../../../prop-types/RoutingErrorShape';
 import ErrorCard from './components/ErrorCard';
-import errorCardProps from './errorCardProperties';
+import findErrorMessageIds from './utils/findErrorMessageIds';
+import errorCardProps from './utils/errorCardProperties';
 
 /**
  * Get error message visual properties.
