+++ app/component/ItinerarySummaryListContainer/components/utils/findErrorMessageIds.js
@@ -1,6 +1,6 @@
 import startsWith from 'lodash/startsWith';
 import distance from '@digitransit-search-util/digitransit-search-util-distance';
-import { PlannerMessageType } from '../../../constants';
+import { PlannerMessageType } from '../../../../constants';
 
 /**
  * @typedef {Object} QueryContext
